requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
black==22.12.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confusable-homoglyphs==3.3.1
crispy-bootstrap5==2024.2
cryptography==43.0.3
cssselect2==0.7.0
Deprecated==1.2.15
Django==5.1.4
django-appconf==1.0.6
django-autoslug==1.9.9
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-imagekit==5.0.0
django-jet==1.0.8
django-jet-reboot==1.3.5
django-model-utils==4.3.1
django-modeltranslation==0.18.11
django-phonenumber-field==8.0.0
django-registration==3.4
django-tables2==2.7.0
djangorestframework==3.15.2
environ==1.0
et-xmlfile==1.1.0
factory_boy==3.3.1
Faker==33.1.0
gopay==2.0.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.3.0
mypy-extensions==1.0.0
mysqlclient==2.2.5
numpy==2.1.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
phonenumbers==8.13.43
pilkit==3.0
pillow==10.4.0
platformdirs==4.3.6
pycparser==2.22
pydantic==1.10.19
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.0.4
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==5.5.0
svglib==1.5.1
tablib==3.6.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.17.0
xhtml2pdf==0.2.15
